fix: truncation of quest ID strings when loading themes quest ID's are 64-bit values so IDSelector must use long, not int
https://github.com/FTBTeam/FTB-Mods-Issues/issues/482

+++ common/src/main/java/dev/ftb/mods/ftbquests/quest/theme/ThemeLoader.java
@@ -161,7 +161,7 @@ public class ThemeLoader implements ResourceManagerReloadListener {
 		}
 
 		try {
-			return new IDSelector(Long.valueOf(sel, 16).intValue());
+			return new IDSelector(Long.valueOf(sel, 16));
 		} catch (Exception ex) {
 			return null;
 		}
